A very important update. MAIN CHANGES: An alternative way of ending the ascension stage has been added for all versions of the algorithm. The ascension stage now automatically ends if the (validation set) accuracy of the network has not changed much recently, even if no layers have settled (their relevance for sources has not reached 1). Specifically, the accuracies measured for the last 'std_window' (default 50) epochs are stored in a list. If, at any epoch, the standard deviation of these 'std_window' accuracies falls below 'std_tolerance' (default 0.1), the algorithm automatically exits the ascension stage. '--std_window' (or '--accuracy_std_window', or '-stdw') and '--std_tolerance' (or '--accuracy_std_tolerance', or '-stdt') can also be manually set as shell arguments.
Added the latest variant of the algorithm (variant 3), which now becomes the default variant.
This is mostly the same as variant 2, except that its ascension stage can only end based on accuracy (through the above mechanism).

The self-constructing algorithm now makes learning rate reductions by default.
The LR reduction protocol to be applied is specified through shell argument '--self_constructing_reduce_lr', or '-rlr' (among other names).
- rlr 0: the learning rate is divided by 10 at 50% or 75% through the improvement stage countdown, but these changes are reverted if the countdown restarts.
- rlr 1: the learning rate is divided by 10 the first time that the countdown reaches 50% or 75%, and these changes are not reverted.
By default rlr 1 is used, since it is the most recent version.

It is now possible to self-structure DenseNet and DenseNet-BC with more than one block.
The algorithm will be used to build each block (once per block), and they will be placed one after the other.
Beforehand, the user must specify how many blocks the network should have.
This is done through the '--block_count' (or '--blocks', or '-bc') shell argument (default value is 1).
Nevertheless, this function is still experimental.

Started to code a variant of the algorithm that features pruning (variant 4), not yet fully implemented.
The idea for now is that the latest block is erased and then rebuilt with the same features (including copied weights), minus the pruned connections.
Most of the relevant functions are found in the 'FOR ADDING BACK PRUNED BLOCKS AND LAYERS' section.
As a corrolary, the code now also keeps a reference to the last block's input channel (because the block could be erased).

OTHER CHANGES:
The '-lnl' parameter for DenseNet-BC does no longer count each layer twice. Before this, -lnl '4' meant 4 composite layers for a DenseNet, but 2 composite layers (and 2 bottlenecks) for a DenseNet-BC.
Changed the SVHN training parameters: max_n_ep = 300, and validation_split = 6000.
Added subsection-marking comments for the 'BUILDING THE DENSENET GRAPH' functions (dense_net.py).
Slightly changed the help string for layer_num_list.
Reflected previous changes in README.